Started installing the void-pointer versions of various routines. I am leaving the original versions as well, to provide backwards compatibility and make testing easier. Once they are no longer needed, I will remove them, or at least clean up their API.

+++ src/t3.h
@@ -3,12 +3,13 @@
 
 #include <stdio.h>
 #include "options.h"
+#include "photon.h"
 #include "types.h"
 
 typedef struct {
 	int32_t channel;
 	int64_t pulse;
-	int64_t time;
+	int32_t time;
 } t3_t;
 
 typedef struct {
